take getActivityData to utils and invoke it from participant-page and participant-list by passing the instance fields to the static method
