+++ src/app/participant-page/participant-page.component.ts
@@ -1226,64 +1226,7 @@ export class ParticipantPageComponent implements OnInit, OnDestroy {
 
   getActivityData(fieldSetting: FieldSettings) {
     //type was activity or activity_staff and no saved staff answer. therefore lookup the activity answer
-    if (fieldSetting != null && fieldSetting.possibleValues != null && fieldSetting.possibleValues[0] != null && fieldSetting.possibleValues[0].value != null) {
-      let tmp: string[] = fieldSetting.possibleValues[ 0 ].value.split( '.' );
-      if (tmp != null && tmp.length > 1) {
-        if (tmp[ 0 ] === 'profile') {
-          return this.participant.data.profile[tmp[1]];
-        }
-        else {
-          if (this.participant != null && this.participant.data != null && this.participant.data.activities != null) {
-            let activity: ActivityData = this.participant.data.activities.find( activity => activity.activityCode === tmp[ 0 ] );
-            if (activity != null && activity.questionsAnswers != null) {
-              let questionAnswer = activity.questionsAnswers.find( questionAnswer => questionAnswer.stableId === tmp[ 1 ] );
-              if (questionAnswer != null) {
-                if (tmp.length == 2) {
-                  if (typeof questionAnswer.answer === "boolean") {
-                    if (questionAnswer.answer) {
-                      return "Yes";
-                    }
-                    return "No";
-                  }
-                  if (questionAnswer.answer instanceof Array) {
-                    return questionAnswer.answer[0];
-                  }
-                  return questionAnswer.answer;
-                }
-                else if (tmp.length === 3) {
-                  if (fieldSetting.possibleValues != null && fieldSetting.possibleValues[ 0 ] != null && fieldSetting.possibleValues[ 0 ].type != null && fieldSetting.possibleValues[ 0 ].type === "RADIO") {
-                    if (questionAnswer.answer != null) {
-                      let found = questionAnswer.answer.find( answer => answer === tmp[ 2 ] )
-                      if (found != null) {
-                        return "Yes";
-                      }
-                      return "No";
-                    }
-                  }
-                  else if (this.activityDefinitions != null) {
-                    let definition: ActivityDefinition = this.activityDefinitions.find( definition => definition.activityCode === tmp[ 0 ] );
-                    if (definition != null && definition.questions != null) {
-                      let question = definition.questions.find( question => question.stableId === tmp[ 1 ] );
-                      if (question != null && question.childQuestions != null) {
-                        for (let i = 0; i < question.childQuestions.length; i++) {
-                          if (question.childQuestions[ i ] != null && question.childQuestions[ i ].stableId === tmp[ 2 ] && questionAnswer.answer[ 0 ][ i ] != null) {
-                            return questionAnswer.answer[ 0 ][ i ];
-                          }
-                        }
-                      }
-                      else if (question != null && question.options != null) {
-
-                      }
-                    }
-                  }
-                }
-              }
-            }
-          }
-        }
-      }
-    }
-    return "";
+    return Utils.getActivityDataValues(fieldSetting, this.participant, this.activityDefinitions);
   }
 
   getActivityOptions(fieldSetting: FieldSettings) {
